+++ packages/alice/lib/core/alice_adapter.dart
@@ -1,9 +1,8 @@
 import 'package:alice/core/alice_core.dart';
 
+/// Adapter mixin which is used in http client adapters.
 mixin AliceAdapter {
   late final AliceCore aliceCore;
 
-  void injectCore(AliceCore aliceCore) {
-    this.aliceCore = aliceCore;
-  }
+  void injectCore(AliceCore aliceCore) => this.aliceCore = aliceCore;
 }