+++ packages/alice/lib/core/alice_core.dart
@@ -1,22 +1,25 @@
-import 'dart:async';
-import 'dart:io';
+import 'dart:async' show StreamSubscription;
+import 'dart:io' show Platform;
 
 import 'package:alice/core/alice_logger.dart';
+import 'package:alice/core/alice_storage.dart';
 import 'package:alice/core/alice_utils.dart';
 import 'package:alice/helper/alice_save_helper.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/model/alice_http_response.dart';
 import 'package:alice/model/alice_log.dart';
-import 'package:alice/ui/page/alice_calls_list_screen.dart';
+import 'package:alice/model/alice_translation.dart';
+import 'package:alice/ui/common/alice_context_ext.dart';
+import 'package:alice/ui/common/alice_navigation.dart';
 import 'package:alice/utils/shake_detector.dart';
-import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:rxdart/rxdart.dart';
+
+typedef AliceOnCallsChanged = Future<void> Function(List<AliceHttpCall>? calls);
 
 class AliceCore {
-  /// Should user be notified with notification if there's new request catched
+  /// Should user be notified with notification if there's new request caught
   /// by Alice
   final bool showNotification;
 
@@ -24,12 +27,22 @@ class AliceCore {
   /// with sensors)
   final bool showInspectorOnShake;
 
-  /// Rx subject which contains all intercepted http calls
-  final BehaviorSubject<List<AliceHttpCall>> callsSubject =
-      BehaviorSubject.seeded([]);
-
   /// Icon url for notification
   final String notificationIcon;
+
+  final AliceStorage _aliceStorage;
+
+  late final NotificationDetails _notificationDetails = NotificationDetails(
+    android: AndroidNotificationDetails(
+      'Alice',
+      'Alice',
+      channelDescription: 'Alice',
+      enableVibration: false,
+      playSound: false,
+      largeIcon: DrawableResourceAndroidBitmap(notificationIcon),
+    ),
+    iOS: const DarwinNotificationDetails(presentSound: false),
+  );
 
   ///Max number of calls that are stored in memory. When count is reached, FIFO
   ///method queue will be used to remove elements.
@@ -43,13 +56,20 @@ class AliceCore {
 
   final AliceLogger _aliceLogger = AliceLogger();
 
-  late FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin;
+  FlutterLocalNotificationsPlugin? _flutterLocalNotificationsPlugin;
+
   GlobalKey<NavigatorState>? navigatorKey;
+
   bool _isInspectorOpened = false;
+
   ShakeDetector? _shakeDetector;
-  StreamSubscription<dynamic>? _callsSubscription;
+
+  StreamSubscription<List<AliceHttpCall>>? _callsSubscription;
+
   String? _notificationMessage;
+
   String? _notificationMessageShown;
+
   bool _notificationProcessing = false;
 
   /// Creates alice core instance
@@ -59,20 +79,19 @@ class AliceCore {
     required this.showInspectorOnShake,
     required this.notificationIcon,
     required this.maxCallsCount,
+    required AliceStorage aliceStorage,
     this.directionality,
     this.showShareButton,
-  }) {
+  }) : _aliceStorage = aliceStorage {
     if (showNotification) {
       _initializeNotificationsPlugin();
       _requestNotificationPermissions();
-      _callsSubscription = callsSubject.listen((_) => _onCallsChanged());
+      _aliceStorage.subscribeToCallChanges(onCallsChanged);
     }
     if (showInspectorOnShake) {
       if (Platform.isAndroid || Platform.isIOS) {
         _shakeDetector = ShakeDetector.autoStart(
-          onPhoneShake: () {
-            navigateToCallListScreen();
-          },
+          onPhoneShake: navigateToCallListScreen,
           shakeThresholdGravity: 4,
         );
       }
@@ -81,35 +100,40 @@ class AliceCore {
 
   /// Dispose subjects and subscriptions
   void dispose() {
-    callsSubject.close();
     _shakeDetector?.stopListening();
     _callsSubscription?.cancel();
   }
 
+  @protected
   void _initializeNotificationsPlugin() {
     _flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
-    final initializationSettingsAndroid =
+    final AndroidInitializationSettings initializationSettingsAndroid =
         AndroidInitializationSettings(notificationIcon);
-    const initializationSettingsIOS = DarwinInitializationSettings();
-    const initializationSettingsMacOS = DarwinInitializationSettings();
-    final initializationSettings = InitializationSettings(
+    const DarwinInitializationSettings initializationSettingsIOS =
+        DarwinInitializationSettings();
+    const DarwinInitializationSettings initializationSettingsMacOS =
+        DarwinInitializationSettings();
+    final InitializationSettings initializationSettings =
+        InitializationSettings(
       android: initializationSettingsAndroid,
       iOS: initializationSettingsIOS,
       macOS: initializationSettingsMacOS,
     );
-    _flutterLocalNotificationsPlugin.initialize(
+    _flutterLocalNotificationsPlugin?.initialize(
       initializationSettings,
       onDidReceiveNotificationResponse: _onDidReceiveNotificationResponse,
     );
   }
 
-  Future<void> _onCallsChanged() async {
-    if (callsSubject.value.isNotEmpty) {
-      _notificationMessage = _getNotificationMessage();
+  @protected
+  Future<void> onCallsChanged(List<AliceHttpCall>? calls) async {
+    if (calls != null && calls.isNotEmpty) {
+      final AliceStats stats = _aliceStorage.getStats();
+
+      _notificationMessage = _getNotificationMessage(stats);
       if (_notificationMessage != _notificationMessageShown &&
           !_notificationProcessing) {
-        await _showLocalNotification();
-        await _onCallsChanged();
+        await _showLocalNotification(stats);
       }
     }
   }
@@ -117,15 +141,13 @@ class AliceCore {
   Future<void> _onDidReceiveNotificationResponse(
     NotificationResponse response,
   ) async {
-    assert(response.payload != null, "payload can't be null");
     navigateToCallListScreen();
-    return;
   }
 
   /// Opens Http calls inspector. This will navigate user to the new fullscreen
   /// page where all listened http calls can be viewed.
   void navigateToCallListScreen() {
-    final context = getContext();
+    final BuildContext? context = getContext();
     if (context == null) {
       AliceUtils.log(
         'Cant start Alice HTTP Inspector. Please add NavigatorKey to your '
@@ -135,86 +157,29 @@ class AliceCore {
     }
     if (!_isInspectorOpened) {
       _isInspectorOpened = true;
-      Navigator.push<void>(
-        context,
-        MaterialPageRoute(
-          builder: (context) => AliceCallsListScreen(this, _aliceLogger),
-        ),
-      ).then((onValue) => _isInspectorOpened = false);
+      AliceNavigation.navigateToCallsList(core: this, logger: _aliceLogger)
+          .then((_) => _isInspectorOpened = false);
     }
   }
 
   /// Get context from navigator key. Used to open inspector route.
   BuildContext? getContext() => navigatorKey?.currentState?.overlay?.context;
 
-  String _getNotificationMessage() {
-    final calls = callsSubject.value;
-    final successCalls = calls
-        .where(
-          (call) =>
-              call.response != null &&
-              call.response!.status! >= 200 &&
-              call.response!.status! < 300,
-        )
-        .toList()
-        .length;
-
-    final redirectCalls = calls
-        .where(
-          (call) =>
-              call.response != null &&
-              call.response!.status! >= 300 &&
-              call.response!.status! < 400,
-        )
-        .toList()
-        .length;
-
-    final errorCalls = calls
-        .where(
-          (call) =>
-              call.response != null &&
-              call.response!.status! >= 400 &&
-              call.response!.status! < 600,
-        )
-        .toList()
-        .length;
-
-    final loadingCalls = calls.where((call) => call.loading).toList().length;
-
-    final notificationsMessage = StringBuffer();
-    if (loadingCalls > 0) {
-      notificationsMessage
-        ..write('Loading: $loadingCalls')
-        ..write(' | ');
-    }
-    if (successCalls > 0) {
-      notificationsMessage
-        ..write('Success: $successCalls')
-        ..write(' | ');
-    }
-    if (redirectCalls > 0) {
-      notificationsMessage
-        ..write('Redirect: $redirectCalls')
-        ..write(' | ');
-    }
-    if (errorCalls > 0) {
-      notificationsMessage.write('Error: $errorCalls');
-    }
-    var notificationMessageString = notificationsMessage.toString();
-    if (notificationMessageString.endsWith(' | ')) {
-      notificationMessageString = notificationMessageString.substring(
-        0,
-        notificationMessageString.length - 3,
-      );
-    }
-
-    return notificationMessageString;
-  }
+  String _getNotificationMessage(AliceStats stats) => <String>[
+        if (stats.loading > 0)
+          '${getContext()?.i18n(AliceTranslationKey.notificationLoading)} ${stats.loading}',
+        if (stats.successes > 0)
+          '${getContext()?.i18n(AliceTranslationKey.notificationSuccess)} ${stats.successes}',
+        if (stats.redirects > 0)
+          '${getContext()?.i18n(AliceTranslationKey.notificationRedirect)} ${stats.redirects}',
+        if (stats.errors > 0)
+          '${getContext()?.i18n(AliceTranslationKey.notificationError)} ${stats.errors}',
+      ].join(' | ');
 
   Future<void> _requestNotificationPermissions() async {
     if (Platform.isIOS || Platform.isMacOS) {
       await _flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
+          ?.resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
           ?.requestPermissions(
             alert: true,
@@ -222,7 +187,7 @@ class AliceCore {
             sound: true,
           );
       await _flutterLocalNotificationsPlugin
-          .resolvePlatformSpecificImplementation<
+          ?.resolvePlatformSpecificImplementation<
               MacOSFlutterLocalNotificationsPlugin>()
           ?.requestPermissions(
             alert: true,
@@ -232,126 +197,72 @@ class AliceCore {
     } else if (Platform.isAndroid) {
       final AndroidFlutterLocalNotificationsPlugin? androidImplementation =
           _flutterLocalNotificationsPlugin
-              .resolvePlatformSpecificImplementation<
+              ?.resolvePlatformSpecificImplementation<
                   AndroidFlutterLocalNotificationsPlugin>();
 
       await androidImplementation?.requestNotificationsPermission();
     }
   }
 
-  Future<void> _showLocalNotification() async {
-    _notificationProcessing = true;
-    const channelId = 'Alice';
-    const channelName = 'Alice';
-    const channelDescription = 'Alice';
-    final androidPlatformChannelSpecifics = AndroidNotificationDetails(
-      channelId,
-      channelName,
-      channelDescription: channelDescription,
-      enableVibration: false,
-      playSound: false,
-      largeIcon: DrawableResourceAndroidBitmap(notificationIcon),
-    );
-    const iOSPlatformChannelSpecifics =
-        DarwinNotificationDetails(presentSound: false);
-    final platformChannelSpecifics = NotificationDetails(
-      android: androidPlatformChannelSpecifics,
-      iOS: iOSPlatformChannelSpecifics,
-    );
-    final message = _notificationMessage;
-    await _flutterLocalNotificationsPlugin.show(
-      0,
-      'Alice (total: ${callsSubject.value.length} requests)',
-      message,
-      platformChannelSpecifics,
-      payload: '',
-    );
+  Future<void> _showLocalNotification(AliceStats stats) async {
+    try {
+      _notificationProcessing = true;
 
-    _notificationMessageShown = message;
-    _notificationProcessing = false;
-    return;
+      final String? message = _notificationMessage;
+
+      await _flutterLocalNotificationsPlugin?.show(
+        0,
+        getContext()
+            ?.i18n(AliceTranslationKey.notificationTotalRequests)
+            .replaceAll("[requestCount]", stats.total.toString()),
+        message,
+        _notificationDetails,
+        payload: '',
+      );
+
+      _notificationMessageShown = message;
+    } finally {
+      _notificationProcessing = false;
+    }
   }
 
   /// Add alice http call to calls subject
-  void addCall(AliceHttpCall call) {
-    final callsCount = callsSubject.value.length;
-    if (callsCount >= maxCallsCount) {
-      final originalCalls = callsSubject.value;
-      final calls = List<AliceHttpCall>.from(originalCalls)
-        ..sort(
-          (call1, call2) => call1.createdTime.compareTo(call2.createdTime),
-        );
-      final indexToReplace = originalCalls.indexOf(calls.first);
-      originalCalls[indexToReplace] = call;
-
-      callsSubject.add(originalCalls);
-    } else {
-      callsSubject.add([...callsSubject.value, call]);
-    }
-  }
+  void addCall(AliceHttpCall call) => _aliceStorage.addCall(call);
 
   /// Add error to existing alice http call
-  void addError(AliceHttpError error, int requestId) {
-    final selectedCall = _selectCall(requestId);
-
-    if (selectedCall == null) {
-      AliceUtils.log('Selected call is null');
-      return;
-    }
-
-    selectedCall.error = error;
-    callsSubject.add([...callsSubject.value]);
-  }
+  void addError(AliceHttpError error, int requestId) =>
+      _aliceStorage.addError(error, requestId);
 
   /// Add response to existing alice http call
-  void addResponse(AliceHttpResponse response, int requestId) {
-    final selectedCall = _selectCall(requestId);
-
-    if (selectedCall == null) {
-      AliceUtils.log('Selected call is null');
-      return;
-    }
-    selectedCall
-      ..loading = false
-      ..response = response
-      ..duration = response.time.millisecondsSinceEpoch -
-          selectedCall.request!.time.millisecondsSinceEpoch;
-
-    callsSubject.add([...callsSubject.value]);
-  }
+  void addResponse(AliceHttpResponse response, int requestId) =>
+      _aliceStorage.addResponse(response, requestId);
 
   /// Add alice http call to calls subject
-  void addHttpCall(AliceHttpCall aliceHttpCall) {
-    assert(aliceHttpCall.request != null, "Http call request can't be null");
-    assert(aliceHttpCall.response != null, "Http call response can't be null");
-    callsSubject.add([...callsSubject.value, aliceHttpCall]);
-  }
+  void addHttpCall(AliceHttpCall aliceHttpCall) =>
+      _aliceStorage.addHttpCall(aliceHttpCall);
 
   /// Remove all calls from calls subject
-  void removeCalls() {
-    callsSubject.add([]);
-  }
+  void removeCalls() => _aliceStorage.removeCalls();
 
-  AliceHttpCall? _selectCall(int requestId) =>
-      callsSubject.value.firstWhereOrNull((call) => call.id == requestId);
+  @protected
+  AliceHttpCall? selectCall(int requestId) =>
+      _aliceStorage.selectCall(requestId);
+
+  Stream<List<AliceHttpCall>> get callsStream => _aliceStorage.callsStream;
+
+  List<AliceHttpCall> getCalls() => _aliceStorage.getCalls();
 
   /// Save all calls to file
   void saveHttpRequests(BuildContext context) {
-    AliceSaveHelper.saveCalls(context, callsSubject.value);
+    AliceSaveHelper.saveCalls(context, _aliceStorage.getCalls());
   }
 
   /// Adds new log to Alice logger.
-  void addLog(AliceLog log) {
-    _aliceLogger.logs.add(log);
-  }
+  void addLog(AliceLog log) => _aliceLogger.logs.add(log);
 
   /// Adds list of logs to Alice logger
-  void addLogs(List<AliceLog> logs) {
-    _aliceLogger.logs.addAll(logs);
-  }
+  void addLogs(List<AliceLog> logs) => _aliceLogger.logs.addAll(logs);
 
   /// Returns flag which determines whether inspector is opened
-  bool isInspectorOpened() {
-    return _isInspectorOpened;
-  }
+  bool get isInspectorOpened => _isInspectorOpened;
 }