+++ packages/alice/lib/helper/alice_conversion_helper.dart
@@ -1,3 +1,4 @@
+/// Helper used in unit conversion.
 class AliceConversionHelper {
   static const int _kilobyteAsByte = 1000;
   static const int _megabyteAsByte = 1000000;
@@ -5,20 +6,15 @@ class AliceConversionHelper {
   static const int _minuteAsMillisecond = 60000;
 
   /// Format bytes text
-  static String formatBytes(int bytes) {
-    if (bytes < 0) {
-      return '-1 B';
-    }
-    if (bytes <= _kilobyteAsByte) {
-      return '$bytes B';
-    }
-    if (bytes <= _megabyteAsByte) {
-      return '${_formatDouble(bytes / _kilobyteAsByte)} kB';
-    }
+  static String formatBytes(int bytes) => switch (bytes) {
+        int bytes when bytes < 0 => '-1 B',
+        int bytes when bytes <= _kilobyteAsByte => '$bytes B',
+        int bytes when bytes <= _megabyteAsByte =>
+          '${_formatDouble(bytes / _kilobyteAsByte)} kB',
+        _ => '${_formatDouble(bytes / _megabyteAsByte)} MB',
+      };
 
-    return '${_formatDouble(bytes / _megabyteAsByte)} MB';
-  }
-
+  /// Formats double with two numbers after dot.
   static String _formatDouble(double value) => value.toStringAsFixed(2);
 
   /// Format time in milliseconds
@@ -33,7 +29,7 @@ class AliceConversionHelper {
       return '${_formatDouble(timeInMillis / _secondAsMillisecond)} s';
     }
 
-    final duration = Duration(milliseconds: timeInMillis);
+    final Duration duration = Duration(milliseconds: timeInMillis);
 
     return '${duration.inMinutes} min ${duration.inSeconds.remainder(60)} s '
         '${duration.inMilliseconds.remainder(1000)} ms';