+++ packages/alice/lib/helper/operating_system.dart
@@ -1,0 +1,9 @@
+/// Definition of OS.
+abstract class OperatingSystem {
+  static const String android = 'android';
+  static const String fuchsia = 'fuchsia';
+  static const String ios = 'ios';
+  static const String linux = 'linux';
+  static const String macos = 'macos';
+  static const String windows = 'windows';
+}