+++ packages/alice_chopper/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 1.0.4
+
+* Create error log on http exception (by Klemen Tusar https://github.com/techouse).
+* Fixed HTTP error status (by Klemen Tusar https://github.com/techouse).
+* Changed linter from very good analysis to flutter lints.
+* General code base refactor.
+
+# 1.0.3
+
+* Replaced milliseconds with UUID for alice token header (by Klemen
+  Tusar https://github.com/techouse).
+* Refactored obtaining the body size (by Klemen Tusar https://github.com/techouse).
+* Removed all unnecessary intermediary variables (by Klemen Tusar https://github.com/techouse).
+
 # 1.0.2
 
 * [BREAKING_CHANGE] Fixed file name of the adapter.